pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "insightsdsa"
version = "0.1.0"
description = "DSA learning insights Flask application"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.0",
    "gunicorn>=22.0",
    "psycopg[binary]>=3.0",
    "psycopg-pool>=3.0",
    "authlib>=1.3",
    "flask-wtf>=1.2",
    "python-dotenv>=1.0",
    "cryptography>=42.0",
    "redis>=5.0",
    "fakeredis>=2.23",
    "requests>=2.31",
    "werkzeug>=3.0",
    "pydantic>=2.0",
    "langchain-core>=0.3",
    "langchain-huggingface>=0.1",
    "langchain-google-genai>=2.0",
    "langchain-openai>=0.2",
    "langchain-xai>=0.1",
    "langgraph>=0.2",
    "langgraph-checkpoint-postgres>=2.0",
    "sqlalchemy>=2.0",
    "pymysql>=1.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "playwright>=1.49.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
insightsdsa = ["templates/**/*", "static/**/*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-q"
